pyproject.toml
[project]
name = "gpu-monitor"
version = "0.1.0"
description = "Basic FastHTML app to show GPU utilization"
readme = "README.md"
# requires-python = ">=3.13"  # TODO
dependencies = [
    "pydantic",
    "python-fasthtml",
    "uvicorn"
]

[tool.ruff]
line-length = 120

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
